Refactor tvdbv2 -> tvdb
