+++ medusa/indexers/config.py
@@ -6,7 +6,7 @@ import re
 
 from medusa.app import BASE_PYMEDUSA_URL
 from medusa.indexers.tmdb.tmdb import Tmdb
-from medusa.indexers.tvdbv2.api import TVDBv2
+from medusa.indexers.tvdb.api import TVDBv2
 from medusa.indexers.tvmaze.api import TVmaze
 from medusa.session.core import MedusaSession
