+++ medusa/metadata/kodi_12plus.py
@@ -11,7 +11,7 @@ from medusa.app import TVDB_API_KEY
 from medusa.helper.common import dateFormat, episode_num
 from medusa.indexers.api import indexerApi
 from medusa.indexers.exceptions import IndexerEpisodeNotFound, IndexerSeasonNotFound
-from medusa.indexers.tvdbv2.api import API_BASE_TVDB
+from medusa.indexers.tvdb.api import API_BASE_TVDB
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.metadata import generic
